% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{addDf2seuratMetaData}
\alias{addDf2seuratMetaData}
\title{Scan Parameters}
\usage{
addDf2seuratMetaData(obj, dfAdd)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
paramerer list
}
\description{
TBD.
}
