% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createDfCoord}
\alias{createDfCoord}
\title{createDfCoord}
\usage{
createDfCoord(obj, params = NULL)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
paramerer list
}
\description{
Scan Parameters
}
