% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{createDfExpr}
\alias{createDfExpr}
\title{Scan Parameters}
\usage{
createDfExpr(obj, assay = "RNA", geneSel = NULL)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
paramerer list
}
\description{
TBD.
}
