% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBwebtools.pckg.r
\name{seurat2viewer}
\alias{seurat2viewer}
\title{Make Seurat Viewer}
\usage{
seurat2viewer(
  obj,
  assay = "RNA",
  geneSel = NULL,
  params = NULL,
  projectName = "test"
)
}
\arguments{
\item{obj}{Seurat object}
}
\value{
app
}
\description{
TBD2.
}
